+++ lib/viewmodels/calendar_view_model.dart
@@ -1,62 +1,44 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-
-final calendarViewModelProvider =
-    StateNotifierProvider<CalendarViewModel, CalendarState>(
-      (ref) => CalendarViewModel(),
-    );
-
-class CalendarState {
-  CalendarState({required this.focusedDay, this.startDay, this.endDay});
-  final DateTime focusedDay;
-  final DateTime? startDay;
-  final DateTime? endDay;
-
-  CalendarState copyWith({
-    DateTime? focusedDay,
-    DateTime? startDay,
-    DateTime? endDay,
-  }) {
-    return CalendarState(
-      focusedDay: focusedDay ?? this.focusedDay,
-      startDay: startDay ?? this.startDay,
-      endDay: endDay ?? this.endDay,
-    );
-  }
-}
+import 'package:travel_muse_app/models/calendar_model.dart';
 
 class CalendarViewModel extends StateNotifier<CalendarState> {
   CalendarViewModel() : super(CalendarState(focusedDay: DateTime.now()));
 
-  // 날짜 선택 처리 (단일 선택 또는 범위 선택)
   void selectDay(DateTime selectedDay, DateTime newFocusedDay) {
     if (state.startDay == null ||
         (state.startDay != null && state.endDay != null)) {
-      // 시작 날짜를 새로 지정하고 종료 날짜 초기화
       state = state.copyWith(
         startDay: selectedDay,
         endDay: null,
         focusedDay: newFocusedDay,
       );
     } else if (state.startDay != null && state.endDay == null) {
-      // 종료 날짜를 지정하거나, 시작 날짜와 위치 바꿈
       if (selectedDay.isBefore(state.startDay!)) {
         state = state.copyWith(
           startDay: selectedDay,
           endDay: state.startDay,
           focusedDay: newFocusedDay,
         );
+      } else if (selectedDay == state.startDay!) {
+        state = state.copyWith(
+          startDay: selectedDay,
+          endDay: null,
+          focusedDay: newFocusedDay,
+        );
       } else {
-        state = state.copyWith(endDay: selectedDay, focusedDay: newFocusedDay);
+        state = state.copyWith(
+          startDay: state.startDay,
+          endDay: selectedDay,
+          focusedDay: newFocusedDay,
+        );
       }
     }
   }
 
-  // 특정 날짜가 선택된 시작일 또는 종료일인지 체크
   bool isSelected(DateTime day) {
     return day == state.startDay || day == state.endDay;
   }
 
-  // 특정 날짜가 시작일과 종료일 사이에 있는지 체크
   bool isBetween(DateTime day) {
     return state.startDay != null &&
         state.endDay != null &&
@@ -64,96 +46,3 @@ class CalendarViewModel extends StateNotifier<CalendarState> {
         day.isBefore(state.endDay!);
   }
 }
-/*
-import 'package:flutter/material.dart';
-import 'package:table_calendar/table_calendar.dart';
-
-class CalendarPage extends StatefulWidget {
-  const CalendarPage({super.key});
-
-  @override
-  _CalendarPageState createState() => _CalendarPageState();
-}
-
-class _CalendarPageState extends State<CalendarPage> {
-  DateTime _focusedDay = DateTime.now();
-  DateTime? _startDay;
-  DateTime? _endDay;
-
-  void _onDaySelected(DateTime selectedDay, DateTime focusedDay) {
-    setState(() {
-      _focusedDay = focusedDay;
-
-      if (_startDay == null || (_startDay != null && _endDay != null)) {
-        _startDay = selectedDay;
-        _endDay = null;
-      } else if (_startDay != null && _endDay == null) {
-        if (selectedDay.isBefore(_startDay!)) {
-          _endDay = _startDay;
-          _startDay = selectedDay;
-        } else {
-          _endDay = selectedDay;
-        }
-      }
-    });
-  }
-
-  bool isBetween(DateTime day) {
-    if (_startDay == null || _endDay == null) return false;
-    return day.isAfter(_startDay!) && day.isBefore(_endDay!);
-  }
-
-  bool isSelected(DateTime day) {
-    return day == _startDay || day == _endDay;
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(title: Text('달력'), centerTitle: true),
-      body: Padding(
-        padding: const EdgeInsets.all(8.0),
-        child: TableCalendar(
-          firstDay: DateTime.utc(2000, 1, 1),
-          lastDay: DateTime.utc(2100, 12, 31),
-          focusedDay: _focusedDay,
-          selectedDayPredicate: isSelected,
-          onDaySelected: _onDaySelected,
-          calendarBuilders: CalendarBuilders(
-            defaultBuilder: (context, day, focusedDay) {
-              final inRange = isBetween(day);
-              return Container(
-                margin: const EdgeInsets.all(4.0),
-                decoration: BoxDecoration(
-                  color: inRange ? Colors.lightBlue.shade100 : null,
-                  borderRadius: BorderRadius.circular(8.0),
-                ),
-                child: Center(child: Text('${day.day}')),
-              );
-            },
-            selectedBuilder: (context, day, focusedDay) {
-              return Container(
-                margin: const EdgeInsets.all(4.0),
-                decoration: BoxDecoration(
-                  color: Colors.blue,
-                  borderRadius: BorderRadius.circular(8.0),
-                ),
-                child: Center(
-                  child: Text(
-                    '${day.day}',
-                    style: TextStyle(
-                      color: Colors.white,
-                      fontWeight: FontWeight.bold,
-                    ),
-                  ),
-                ),
-              );
-            },
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-*/
