+++ lib/views/calendar/calendar_page.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:travel_muse_app/viewmodels/calendar_view_model.dart';
+import 'package:travel_muse_app/providers/calendar_provider.dart';
 import 'package:travel_muse_app/views/calendar/widgets/calendar_header.dart';
 import 'package:travel_muse_app/views/calendar/widgets/calendar_widget.dart';
